requirements.txt
Django==2.0.0
Pillow==5
djangorestframework==3.8.0
# djangorestframework-jwt abondaned no latest django support
djangorestframework_simplejwt
pytest==4.0.1
pytest-django==3.4.4
arrow==0.11
requests==2.20
# psycopg2
psycopg2-binary==2.7.6